fritz_to_influx.py script
This is the initial commit of the working fritz_to_influx.py script.

Running this program will connect to a Fritz!BOX and extract data from
it every 20 seconds.  This data will then be sent to an Influx database
using the 'section' and 'action' from the data as tags and the property
names as field keys.  The locationa dn credentials of both Fritz!BOX
and Influx database can be controlled via environment variables, which
allows the Fritz!BOX password to be hidden from view.

Signed-off-by: Paul Wayper <paulway@mabula.net>